third_party/WebKit/LayoutTests/fast/dom/geometry-interfaces-dom-point.html

<!doctype html>
<meta charset="utf-8">
<title>Geometry Interfaces: DOMPoint</title>
<link rel="help" href="https://drafts.fxtf.org/geometry/#DOMPoint">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/geometry-interfaces-test-helpers.js"></script>
<script>
'use strict';

test(() => {
  var point = new DOMPoint(2, 3);
  assert_dom_point_equals(point, [2, 3, 0, 1]);
}, 'DOMPoint constructor with x, y parameters');

test(() => {
  var point = new DOMPoint(5, 7, 9);
  assert_dom_point_equals(point, [5, 7, 9, 1]);
}, 'DOMPoint constructor with x, y, z parameters');

test(() => {
  var point = new DOMPoint(8, 2, 1, 6);
  assert_dom_point_equals(point, [8, 2, 1, 6]);
}, 'DOMPoint constructor with x, y, z, w parameters');

test(() => {
  var point = new DOMPoint({x: 2});
  assert_dom_point_equals(point, [2, 0, 0, 1]);
}, 'DOMPoint constructor with x parameter');

test(() => {
  var point = new DOMPoint({y: 2});
  assert_dom_point_equals(point, [0, 2, 0, 1]);
}, 'DOMPoint constructor with y parameter');

test(() => {
  var point = new DOMPoint({z: 2});
  assert_dom_point_equals(point, [0, 0, 2, 1]);
}, 'DOMPoint constructor with z parameter');

test(() => {
  var point = new DOMPoint({w: 2});
  assert_dom_point_equals(point, [0, 0, 0, 2]);
}, 'DOMPoint constructor with w parameter');

test(() => {
  var point = new DOMPoint({x: 2, y: 3, z: 4, w: 5});
  assert_dom_point_equals(point, [2, 3, 4, 5]);
}, 'DOMPoint constructor with x, y, z, w parameters');

test(() => {
  var point = new DOMPoint();
  assert_dom_point_equals(point, [0, 0, 0, 1]);
}, 'DOMPoint constructor without parameter');

test(() => {
  var point = new DOMPoint();
  point.x = 10;
  point.y = 20;
  point.z = 30;
  point.w = 40;
  assert_dom_point_equals(point, [10, 20, 30, 40]);
}, 'DOMPoint setter');

test(() => {
  var point = new DOMPoint(1, 2, 3, 4);
  assert_object_equals(point.toJSON(), {x: 1, y: 2, z: 3, w: 4});
}, 'DOMPoint toJSON');

</script>
